+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+# [1.15.0](https://github.com/wolasss/random-scale-trainer/compare/v1.14.0...v1.15.0) (2026-08-20)
+
+
+### Features
+
+* **scoring:** bonus for playing the note in time with the tick ([#127](https://github.com/wolasss/random-scale-trainer/issues/127)) ([7a82344](https://github.com/wolasss/random-scale-trainer/commit/7a82344d7af5be8648a9bc809c17c59e64a2b42b))
+
+# [1.14.0](https://github.com/wolasss/random-scale-trainer/compare/v1.13.0...v1.14.0) (2026-08-20)
+
+
+### Features
+
+* **scoring:** bonus for playing the called note in two octaves ([#124](https://github.com/wolasss/random-scale-trainer/issues/124)) ([a10f00d](https://github.com/wolasss/random-scale-trainer/commit/a10f00d7457cd6c387479e85075b4beeb3ae7c86))
+
 # [1.13.0](https://github.com/wolasss/random-scale-trainer/compare/v1.12.3...v1.13.0) (2026-08-19)